chore: move all request and response structs to types package and update references in request functions accordingly
