+++ platform/types/asset_group.go
